pyproject.toml
[project]
name = "geotag"
version = "0.1.0"
description = "Photo geotagging application with web GUI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pandas>=2.2.0",
    "pillow>=11.0.0",
    "gpxpy>=1.6.0",
    "python-multipart>=0.0.12",
    "requests>=2.32.0",
    "pyyaml>=6.0.0",
    "piexif>=1.1.3",
    "iptcinfo3>=2.1.4",
    "pywin32>=306; sys_platform == 'win32'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
